apps/api/requirements.txt
flask==3.1.0
flask-restful==0.3.10
flask-cors==5.0.0
flask-jwt-extended==4.7.1
marshmallow==3.23.1
supabase==2.11.0
python-dotenv==1.0.1
gunicorn==23.0.0
sentry-sdk[flask]==2.19.2
